requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.5; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.5.0; python_version >= '3.7'
anyio==3.7.1; python_version >= '3.7'
async-timeout==4.0.3; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
black==23.9.1; python_version >= '3.8'
blis==0.7.11
catalogue==2.0.10; python_version >= '3.6'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cloudpathlib==0.15.1; python_version >= '3.7'
confection==0.1.3; python_version >= '3.6'
cymem==2.0.8
exceptiongroup==1.1.3; python_version < '3.11'
fastapi==0.103.2; python_version >= '3.7'
frozenlist==1.4.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
isort==5.12.0; python_full_version >= '3.8.0'
jinja2==3.1.2; python_version >= '3.7'
joblib==1.3.2; python_version >= '3.7'
langcodes==3.3.0; python_version >= '3.6'
markupsafe==2.1.3; python_version >= '3.7'
metaphor-python==0.1.16
multidict==6.0.4; python_version >= '3.7'
murmurhash==1.0.10; python_version >= '3.6'
mypy-extensions==1.0.0; python_version >= '3.5'
nltk==3.8.1; python_version >= '3.7'
numpy==1.24.4; python_version >= '3.8'
openai==0.28.1; python_full_version >= '3.7.1'
packaging==23.2; python_version >= '3.7'
pathspec==0.11.2; python_version >= '3.7'
pathy==0.10.2; python_version >= '3.6'
platformdirs==3.11.0; python_version >= '3.7'
preshed==3.0.9; python_version >= '3.6'
pydantic==2.4.2; python_version >= '3.7'
pydantic-core==2.10.1; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
regex==2023.8.8; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
setuptools==68.2.2; python_version >= '3.8'
smart-open==6.4.0; python_version >= '3.6' and python_version < '4.0'
sniffio==1.3.0; python_version >= '3.7'
spacy[en_core_web_sm]==3.7.0; python_version >= '3.7'
spacy-legacy==3.0.12; python_version >= '3.6'
spacy-loggers==1.0.5; python_version >= '3.6'
srsly==2.4.8; python_version >= '3.6'
starlette==0.27.0; python_version >= '3.7'
thinc==8.2.1; python_version >= '3.6'
tiktoken==0.5.1; python_version >= '3.8'
tomli==2.0.1; python_version < '3.11'
tqdm==4.66.1; python_version >= '3.7'
typer==0.7.0; python_version >= '3.6'
typing-extensions==4.8.0; python_version < '3.11'
urllib3==2.0.6; python_version >= '3.7'
uvicorn==0.23.2; python_version >= '3.8'
wasabi==1.1.2; python_version >= '3.6'
weasel==0.3.1; python_version >= '3.6'
yarl==1.9.2; python_version >= '3.7'
